Rewriting: two more approaches to rewrite rules
Both implement rewrite rules as procedures of type (Term -> [Term]),
where an empty list signals failure, a singleton list contains the
result, and a list of multiple elements is currently unused. In the
future we could imagine implementing term matching in a way which
returns multiple results, for example matching terms modulo
commutativity.

Encoding rewrite rules as procedures is convenient when using them, but
has the disadvantage that they cannot be easily manipulated after
definition. For example, if we wanted to flip a rewrite rule so it goes
right to left instead of left to right, there is no obvious way to do
this when it's encoded as a procedure of type (Term -> [Term])

The second approach implements the rewrite rules using Racket's bundled
match functionality. This has the disadvantage that we are not able to
modify the matching algorithm, for example matching modulo commutativity
as mentioned above. But as an advantage, it provides a lot of free
functionality.