Jupyter Lab + Wolfram Engine
Setup for experimenting with the Wolfram Language, which is the only
mainstream rewriting based language I'm currently aware of.

Wolfram Engine wants to run its install script as root for some
ridiculous reason, which is obviously not going to fly on my machine so
I've gone with the docker approach.